feat: separate 'edit'
